requirements.txt
numpy
pandas
gsw
trollsift
six
PyYAML
setuptools
